web/api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.0
pydantic>=2.7.0
pyyaml>=6.0
